refactor : 공통 요소들 묶고 컴포넌트 분리
