+++ client/user/src/app/(nonRoot)/(auth)/(find)/id/page.tsx
@@ -1,0 +1,5 @@
+import { AutoRedirect } from '@tookscan/components'
+
+export default function Id() {
+  return <AutoRedirect to="/id/reset" />
+}
