--- client/user/src/app/(nonRoot)/(auth)/id/page.tsx
@@ -1,7 +1,0 @@
-'use client'
-
-import { AutoRedirect } from '@tookscan/components/app/AutoRedirect'
-
-export default function Id() {
-  return <AutoRedirect to="/id/reset" />
-}
