+++ packages/components/ui/Modal/index.ts
@@ -1,3 +1,4 @@
+export * from './Confirm'
 export * from './Modal'
 export * from './SearchAddress'
 export * from './Toast'
